Set cache-timeout to max when building local playlists
This should be fine since no time sensitive information is needed in these requests, just like the title and author info, and such.

+++ src/helpers.rs
@@ -141,7 +141,7 @@ pub async fn get_previous_data(collection: &str, key: &str, db: &DbWrapper, app_
           Some(timestamp) => {
             let current_timestamp = Utc::now().timestamp();
             let offset = current_timestamp - timestamp;
-            if offset as i32 > app_settings.cache_timeout {
+            if offset as u64 > app_settings.cache_timeout {
               db.delete(collection, key).await;
               None
             } else {
